% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyReg.R
\name{applyReg}
\alias{applyReg}
\title{Apply already computed time warping to a set of curves}
\usage{
applyReg(dat, reg, grid = NULL, id = NULL, time = NULL, value = NULL)
}
\arguments{
\item{dat}{Either a long form dataframe or a (\code{*})\code{funData} object containing a data set of curves.
In the dataframe case, it should have at least three columns mapping to \code{id} (the curve ID),
\code{time} (the time sample) and \code{value} (the corresponding value).
The dataframe is assumed to be ordered by \code{id} and within \code{id} by \code{time}.
The order of \code{id} should match the order of h(t) in \code{reg}.
In the case of (\code{*})\code{funData}, the order of stored curves (in \verb{@X}) should match the order of h(t) in \code{reg}.}

\item{reg}{A list with at least key \code{h}, containing the result of landmark registration operated by \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}}.}

\item{grid}{A vector of time samples defined on the registered time axis.
If specified, the value of the curves in \code{dat} at those time points are computed,
which entails an interpolation of the original curves (via \code{fda::Data2fd()}).
If \code{NULL}, the registered curve values will be computed at the respective h(t),
i.e. at the time points on the registered time axis corresponding to those provided in \code{dat}.
In the latter case, \code{reg} should include the \code{hinv} entry.}

\item{id}{Column name of \code{dat} corresponding to curve index (default: 1st column).
Not used if \code{dat} is a (\code{*})\code{funData}.}

\item{time}{Column name of \code{dat} corresponding to the time axis (default: 2nd column).
Not used if \code{dat} is a (\code{*})\code{funData}.}

\item{value}{Column name(s) of \code{dat} corresponding to the values at \code{time}.
For multi-dimensional data, a vector of strings (default: columns from 3rd to last).
Not used if \code{dat} is a (\code{*})\code{funData}.}
}
\value{
The registered curves in the same format as \code{dat}.
}
\description{
Given a set of curves and a corresponding set of time warping functions h(t),
typically computed with \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}}, obtain the transformed curves.
The curves can be encoded either as long form dataframe or as (\code{irreg})\code{funData}.
The result is provided in the same format as the input.
A desired time point grid on the registered time axis can be provided,
otherwise the time points for each curve are obtained by mapping the original ones according to the respective h(t).
}
\examples{
library(dplyr)
library(tibble)
# first generate a landmark registration object
inputMarks <- matrix(c(0, 0.5, 1, 2, 0, 0.7, 1.2, 1.9, 0, 0.4, 1.1, 2.2), nrow=3, byrow = TRUE)
reg <- landmarkreg_nocurves(inputMarks)
# generate corresponding curves
curves <- tibble(id = 1:3) \%>\%
  group_by(id) \%>\%
  reframe(time = seq(0, max(inputMarks[id,]), by = 0.1), value = 0.1 * id **2)
# version without specification of grid
regCurves <- applyReg(curves, reg)
# version with grid
grid <- seq(0, 2, by = 0.5)
regCurves <- applyReg(curves, reg, grid)
}
