% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyRegDf.R
\name{applyRegDf}
\alias{applyRegDf}
\title{Apply already computed time warping to a set of uni-dimensional curves in dataframe long format}
\usage{
applyRegDf(df, reg, grid = NULL, id, time, value)
}
\arguments{
\item{df}{A dataframe containing at least three columns}

\item{reg}{A list with at least key \code{h}, containing the result of landmark registration operated by \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}}.}

\item{grid}{A vector of time samples defined on the registered time axis.
If specified, the value of the curves in \code{df} at those time points will be computed,
which entails an interpolation of the original curves (via \code{fda::Data2fd()}).
If \code{NULL}, the registered curve values will be computed at the respective h(t),
i.e. at the time points on the registered time axis corresponding to those provided in \code{df}.
In the latter case, \code{reg} should include the \code{hinv} entry.}

\item{id}{Column name of \code{df} corresponding to curve index}

\item{time}{Column name of \code{df} corresponding to the time axis}

\item{value}{Column name of \code{df} corresponding to the values at \code{time}}
}
\value{
A tibble containing the registered curves.
The tibble has three columns named as specified by \code{id}, \code{time} and \code{value}.
If \code{grid} is \code{NULL}, the tibble has the same number of rows as \code{df},
its \code{id} and \code{value} columns are the same as their namesake in \code{df},
the \code{time} column contains the warped time values corresponding to their namesake in \code{df},
i.e. hinv(\code{time}).
If \code{grid} is specified, each curve is evaluated at \code{time} points specified by \code{grid}
(hence the tibble has \code{length(grid) * length(unique(id))} rows and three columns).
}
\description{
Given a set of curves and a corresponding set of time warping functions,
typically computed with \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}}, obtain the transformed curves.
The curves are encoded in long form dataframe.
The result is provided in the same format.
A desired time point grid on the registered time axis can be provided,
otherwise the time points for each curve are obtained by mapping the original ones according to the respective h(t).
}
\details{
This function does not check its arguments nor it provides default values.
The user is advised to use \code{\link[=applyReg]{applyReg()}} instead, which calls this function internally.
}
