% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defint_fd.R
\name{defint_fd}
\alias{defint_fd}
\title{Definite integral}
\usage{
defint_fd(fdobj, rng = NULL)
}
\arguments{
\item{fdobj}{Object of class \code{fda::fd} containing uni-dimensional curve(s) to integrate}

\item{rng}{Numeric vector of length 2 specifying the limits of integration.
If \code{NULL}, the entire range of \code{fdobj} is used.}
}
\value{
A numerical vector of length = number of functions in \code{fdobj}
containing the integral values.
}
\description{
Compute a definite integral for each of the uni-dimensional functions
in \code{fdobj} from \code{rng[1]} to \code{rng[2]}.
}
