% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funData2long.R
\name{funData2long}
\alias{funData2long}
\title{Convert a (irreg)FunData object into a dataframe}
\usage{
funData2long(fd, columnNames = NULL)
}
\arguments{
\item{fd}{A (irreg)FunData object}

\item{columnNames}{(optional) list indicating the column names corresponding to \code{fd@argvals}, \code{fd@X},
and the curve ID, indicated by \code{.index}, \code{.value} and \code{.obs}, respectively.
If \code{NULL}, column names are \code{argvals}, \code{X} and \code{ID}.}
}
\value{
A three-column tibble in long format
}
\description{
Convert a (irreg)FunData object into a dataframe
}
