% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkreg_nocurves.R
\name{landmarkreg_nocurves}
\alias{landmarkreg_nocurves}
\title{Compute several landmark-based time warping function}
\usage{
landmarkreg_nocurves(
  inputMarks,
  targetMarks = NULL,
  compute_hinv = FALSE,
  njobs = 1,
  WfdPar = NULL,
  wlambda = 1e-14
)
}
\arguments{
\item{inputMarks}{Matrix or dataframe, each row obeys the same conventions as the namesake
argument in \code{one_landmarkreg_nocurves}}

\item{targetMarks}{Same as in \code{one_landmarkreg_nocurves}, all \code{inputMarks} have the same target.
If \code{NULL}, it is set as the mean by column of \code{inputMarks}.}

\item{compute_hinv}{If \code{TRUE} return hinv, the inverse function of h(t)}

\item{njobs}{If > 1, use a parallel procedure}

\item{WfdPar}{(optional, see \code{one_landmarkreg_nocurves})}

\item{wlambda}{(optional, see \code{one_landmarkreg_nocurves})}
}
\value{
A list with the following keys:

\code{h}: a \code{fd} object with as many h(t) as the number of rows of \code{inputMarks}

\code{lograte}: - log (dh/dt)

\code{hinv}: (optional) a list of \code{fd} objects, one for each row of \code{inputMarks}
}
\description{
Applies \code{one_landmarkreg_nocurves} to each row of \code{inputMarks}.
Returns all h(t) and log rates in \code{fda::fd} form.
The number of basis for these \code{fd} objects is very large to allow precise interpolation
and avoid having spurious negative values in dh/dt at any point in time.
Anyway h is not strictly guaranteed monotonic
(contrary to its samples coming from \code{one_landmarkreg_nocurves})
and its extremes are not precise.
If \code{compute_hinv=TRUE}, returns also all h_inv(t) (TODO: as well as a matrix of errors
as difference between \code{targetMarks} and the actual registered positions).
}
\examples{
inputMarks <- matrix(c(0, 0.5, 1, 2, 0, 0.7, 1.2, 1.9, 0, 0.4, 1.1, 2.2), nrow=3, byrow = TRUE)
reg <- landmarkreg_nocurves(inputMarks)
}
