% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkreg_timeSamples.R
\name{landmarkreg_timeSamples}
\alias{landmarkreg_timeSamples}
\title{Landmark registration directly on time samples}
\usage{
landmarkreg_timeSamples(
  timeSamples,
  inputMarks,
  targetMarks,
  WfdPar = NULL,
  wlambda = 1e-14
)
}
\arguments{
\item{timeSamples}{Vector (not matrix) of time input samples (e.g. regularly spaced)}

\item{inputMarks}{Numeric vectors of increasing time values, starting with 0.}

\item{targetMarks}{Numeric vectors of increasing time values, starting with 0.}

\item{WfdPar}{(optional, see \code{one_landmarkreg_nocurves})}

\item{wlambda}{(optional, see \code{one_landmarkreg_nocurves})}
}
\value{
h(\code{timeSamples}), i.e. a vector of the same length as \code{timeSamples} with time-registered values
corresponding to \code{timeSamples}
}
\description{
Landmark registration applied directly to the input time axis samples.
Both input and target landmarks must start from 0.
}
\examples{
t <- seq(0,2, by=0.1)
x <- t **2 - 0.5 * t
inputMarks <- c(0, 1, 1.5, 2)
targetMarks <- c(0, 0.5, 1, 1.3)
t_reg <- landmarkreg_timeSamples(t, inputMarks, targetMarks)
plot(t, x, type="b", pch=19, col="red")
lines(t_reg, x, pch=18, col="blue", type="b", lty=2, xlab="t")
legend(0, 3, legend = c("original", "registered"), col=c("red", "blue"), lty=1:2, cex=0.8)
}
