% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks2durations.R
\name{landmarks2durations}
\alias{landmarks2durations}
\title{Convert landmark positions to interval durations}
\usage{
landmarks2durations(inputMarks, id = NULL, targetMarks = NULL)
}
\arguments{
\item{inputMarks}{A dataframe with landmark information}

\item{id}{(optional) column namr of \code{inputMarks} that contains curve index}

\item{targetMarks}{(optional) The position of landmarks after registration}
}
\value{
A long format dataframe of intervals.
}
\description{
Given a wide format dataframe listing landmark positions, one row per curve,
convert into a long format dataframe of intervals.
}
\examples{
library(tibble)
land <- tribble(
~curveId, ~l1, ~l2, ~l3, ~l4,
1, 0, 0.7, 1, 1.5,
2, 0, 0.5, 1.1, 1.8,
3, 0, 0.6, 1.2, 2
)
landmarks2durations(land, "curveId")
}
