% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lograte2duration.R
\name{lograte2duration}
\alias{lograte2duration}
\title{Convert registered interval to its original duration}
\usage{
lograte2duration(lograte, from, to)
}
\arguments{
\item{lograte}{Either a \code{fda::fd} or a \code{funData::funData} object containing only one curve
which represents the log rage to be applied}

\item{from}{The lower interval boundary on the registered time axis}

\item{to}{The upper interval boundary on the registered time axis}
}
\value{
The duration of the registered interval (\code{from}, \code{to}) on the original time axis.
}
\description{
Given a log rate function, computed with \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}},
convert a time interval (\code{from}, \code{to}) specified on the registered time axis
back to its original duration
}
