% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2irregFunData.R
\name{long2irregFunData}
\alias{long2irregFunData}
\title{Convert a set of curves encoded as long form dataframe to a irregFunData}
\usage{
long2irregFunData(df, columnNames)
}
\arguments{
\item{df}{A dataframe containing at least three columns}

\item{columnNames}{A list with three entries: \code{.obs}, \code{.index} and \code{.value},
the values of which are the column names of \code{df} corresponding to
curve ID, time samples and sample values, respectively.}
}
\value{
An object of class \code{funData::irregFunData}
}
\description{
Convert a set of curves encoded as long form dataframe to a irregFunData
}
