% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2list.R
\name{long2list}
\alias{long2list}
\title{Convert a two-column long form dataframe into a list}
\usage{
long2list(dat, keys, values)
}
\arguments{
\item{dat}{A dataframe in long format containing at least two columns}

\item{keys}{String, the column name that will become the list key}

\item{values}{String, the column name that will become the value}
}
\value{
A list with as many keys as there are unique values of the \code{keys} column,
each key containing a vector of the corresponding values in the \code{values} column.
}
\description{
Convert a two-column long form dataframe into a list
}
