% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_landmarkreg_nocurves.R
\name{one_landmarkreg_nocurves}
\alias{one_landmarkreg_nocurves}
\title{Compute one landmark-based time warping function}
\usage{
one_landmarkreg_nocurves(
  inputMarks,
  targetMarks,
  WfdPar = NULL,
  wlambda = 1e-14
)
}
\arguments{
\item{inputMarks}{Numeric vectors of increasing time values, starting with 0.}

\item{targetMarks}{Numeric vectors of increasing time values, starting with 0.}

\item{WfdPar}{(optional) fdPar object to compute the warping, it should be defined on the (0,1) range}

\item{wlambda}{(optional) lambda for the construction of WfdPar}
}
\value{
A vector of samples of h(t), whose t values are (implicitly) given as regularly spaced
between 0 and the last value in targetMarks.
The first value of h(t) is 0, the last is the same as the last value in inputMarks.
The number of samples is large enough to provide for a smooth interpolation
operated by calling functions like landmarkreg_nocurve.
}
\description{
Computes the time warping function h(t) so that input landmarks are moved to corresponding target landmarks.
Both input and target landmarks are first linearly mapped to (0,1),
then the code computes h(t) samples using the same procedure used in fda::landmarkreg.
Both input and target landmarks must start from 0.
The scaling to (0,1) (and reverse scaling at the end) is necessary
as smooth.morph does not yield the expected results
when operating on other intervals, e.g. when they differ too much from one another.
}
\examples{
inputMarks <- c(0, 1, 1.5, 2)
targetMarks <- c(0, 0.5, 1, 1.3)
h <- one_landmarkreg_nocurves(inputMarks, targetMarks)
plot(h, type = 'l', xlab = 'registered time', ylab = 'original time')
}
