% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_at_treg.R
\name{val_at_treg}
\alias{val_at_treg}
\title{Compute values of a curve corresponding to points on a time-warped axis}
\usage{
val_at_treg(argvals, vals, argvalreg, h)
}
\arguments{
\item{argvals}{A vector of original time points}

\item{vals}{A vector of values corresponding to \code{argvals}}

\item{argvalreg}{A vector of time points on the warped time axis,
unrelated to \code{argvals} either in number or values}

\item{h}{A \code{fda::fd} object containing only one curve}
}
\value{
A vector of curve values of the same length as \code{argvalreg}
}
\description{
Given a curve defined by (\code{argvals}, \code{vals}) pairs, and given an already computed
time warping function \code{h} (mapping registered time to original time),
compute the values of the curve at time points \code{argvalreg} on the time warped axis.
The computation entails an interpolation of the original curve via \code{fda::Data2fd()}.
}
\details{
The range of \code{argvals} implicitly defines the expected range (co-domain) of \code{h}.
In case the min or max values of \code{h}(\code{argvalreg}) exceed the range of \code{h}
(because of rounding errors produced in \code{\link[=landmarkreg_nocurves]{landmarkreg_nocurves()}}),
these extremes are substituted into \code{argvals}.
This is an attempt to prevent common crashes due to approximations in \code{h},
typically when the value \code{h}(0) is slightly below zero.
}
