% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WfdPar_for_reg.R
\name{WfdPar_for_reg}
\alias{WfdPar_for_reg}
\title{Wrapper to fdPar}
\usage{
WfdPar_for_reg(marks, lambda = 1e-10)
}
\arguments{
\item{marks}{A vector of strictly increasing time points}

\item{lambda}{Roughness penalty weight}
}
\value{
An object of class fda::fdPar
}
\description{
A helper function that builds a fda::fdPar object given a set of time points.
It is used by one_landmarkreg_nocurves to pass an fdPar object to smooth.morph
}
\examples{
targetMarks <- c(0, 1.2, 2, 3.4)
fdParObj <- WfdPar_for_reg(targetMarks)
}
