% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funData2long1.R
\name{funData2long1}
\alias{funData2long1}
\title{Convert a (irreg)FunData object containing only one curve into a dataframe}
\usage{
funData2long1(fd, columnNames = NULL)
}
\arguments{
\item{fd}{A (irreg)FunData object containing only one curve}

\item{columnNames}{(optional) list indicating the column names corresponding to \code{fd@argvals} and \code{fd@X},
indicated by \code{.index} and \code{.value}, respectively.
If \code{NULL}, column names are \code{argvals} and \code{X}.}
}
\value{
A two-column tibble in long format
}
\description{
Convert a (irreg)FunData object containing only one curve into a dataframe
}
