% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2irregFunData.R
\name{long2irregFunData}
\alias{long2irregFunData}
\title{Convert a set of curves encoded as long form dataframe to an irregFunData object}
\usage{
long2irregFunData(
  df,
  id = names(df)[1],
  time = names(df)[2],
  value = names(df)[3]
)
}
\arguments{
\item{df}{A dataframe containing at least three columns}

\item{id}{Column name of \code{df} corresponding to curve index (default: 1st column of \code{df})}

\item{time}{Column name of \code{df} corresponding to the time axis (default: 2nd column of \code{df})}

\item{value}{Column name of \code{df} corresponding to the values at \code{time} (default: 3rd column of \code{df})}
}
\value{
An object of class \code{funData::irregFunData}
}
\description{
The long form dataframe should have at least three columns: curve id, time and value.
Their names have to be specified, else they are assumed to be the first three columns in this order: id, time, value.
Time and value specify samples for each curve, curve id identifies the curves.
The conversion to \code{funData::irregFunData} places the first curve at \code{argvals[[1]]}, \code{X[[1]]},
the second curve at \code{argvals[[2]]}, \code{X[[2]]}, etc., using the order of appearance of curve id in the dataframe,
i.e. not alphabetic or factor order.
Note that the original id's are not stored in the \code{irregFunData} object.
}
